# app/backend/contracts/openapi.v0.1.json
{
  "components": {
    "schemas": {
...
      "CreateReportAssistantDraftRequest": {
        "additionalProperties": false,
        "properties": {
          "artifact_id": {
            "format": "uuid",
            "title": "Artifact Id",
            "type": "string"
          },
          "instruction": {
            "default": "경영 검토용 보고서 초안을 구성해 줘",
            "maxLength": 500,
            "minLength": 1,
            "title": "Instruction",
            "type": "string"
          }
        },
        "required": [
          "artifact_id"
        ],
        "title": "CreateReportAssistantDraftRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ReportAssistantDraftResponse": {
        "additionalProperties": false,
        "properties": {
          "assistant_request_id": {
            "format": "uuid",
            "title": "Assistant Request Id",
            "type": "string"
          },
          "definition": {
            "$ref": "#/components/schemas/ReportDefinitionResponse"
          },
          "status": {
            "const": "success",
            "title": "Status",
            "type": "string"
          },
          "trace": {
            "$ref": "#/components/schemas/ReportAssistantTraceResponse"
          }
        },
        "required": [
          "assistant_request_id",
          "status",
          "definition",
          "trace"
        ],
        "title": "ReportAssistantDraftResponse",
        "type": "object"
      },
...
      "ReportDefinitionListResponse": {
        "additionalProperties": false,
        "properties": {
          "contract_version": {
            "title": "Contract Version",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ReportDefinitionResponse"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "contract_version",
          "items"
        ],
        "title": "ReportDefinitionListResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "description": "서버가 AUTH_PRINCIPALS_FILE의 SHA-256 digest로 검증하는 Bearer token",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Answervice Control Plane",
    "version": "OPENAPI-v1.1.0-DRAFT"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/reports/assistant/drafts": {
      "post": {
        "operationId": "reportAssistantCreateDraft",
        "parameters": [
          {
            "in": "header",
            "name": "X-As-Of",
            "required": true,
            "schema": {
              "title": "X-As-Of",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Trace-Id",
            "required": true,
            "schema": {
              "title": "X-Trace-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Timezone",
            "required": true,
            "schema": {
              "title": "X-Timezone",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Contract-Version",
            "required": true,
            "schema": {
              "title": "X-Contract-Version",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReportAssistantDraftRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportAssistantDraftResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Create Assistant Draft"
      }
    },
    "/reports/definitions": {
      "get": {
        "operationId": "reportListDefinitions",
        "parameters": [
          {
            "in": "header",
            "name": "X-As-Of",
            "required": true,
            "schema": {
              "title": "X-As-Of",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Trace-Id",
            "required": true,
            "schema": {
              "title": "X-Trace-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Timezone",
            "required": true,
            "schema": {
              "title": "X-Timezone",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Contract-Version",
            "required": true,
            "schema": {
              "title": "X-Contract-Version",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportDefinitionListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "List Definitions"
      },
      "post": {
        "operationId": "reportCreateDefinition",
        "parameters": [
          {
            "in": "header",
            "name": "X-As-Of",
            "required": true,
            "schema": {
              "title": "X-As-Of",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Trace-Id",
            "required": true,
            "schema": {
              "title": "X-Trace-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Timezone",
            "required": true,
            "schema": {
              "title": "X-Timezone",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Contract-Version",
            "required": true,
            "schema": {
              "title": "X-Contract-Version",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReportDefinitionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportDefinitionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Create Definition"
      }
    },
    "/reports/definitions/{definition_id}/versions/{version}": {
      "get": {
        "operationId": "reportGetDefinitionVersion",
        "parameters": [
          {
            "in": "path",
            "name": "definition_id",
            "required": true,
            "schema": {
              "title": "Definition Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "version",
            "required": true,
            "schema": {
              "title": "Version",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-As-Of",
            "required": true,
            "schema": {
              "title": "X-As-Of",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Trace-Id",
            "required": true,
            "schema": {
              "title": "X-Trace-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Timezone",
            "required": true,
            "schema": {
              "title": "X-Timezone",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Contract-Version",
            "required": true,
            "schema": {
              "title": "X-Contract-Version",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportDefinitionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Get Version"
      }
    },
    "/reports/definitions/{definition_id}/versions/{version}/approve": {
      "post": {
        "operationId": "reportApproveVersion",
        "parameters": [
          {
            "in": "path",
            "name": "definition_id",
            "required": true,
            "schema": {
              "title": "Definition Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "version",
            "required": true,
            "schema": {
              "title": "Version",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-As-Of",
            "required": true,
            "schema": {
              "title": "X-As-Of",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Trace-Id",
            "required": true,
            "schema": {
              "title": "X-Trace-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Timezone",
            "required": true,
            "schema": {
              "title": "X-Timezone",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Contract-Version",
            "required": true,
            "schema": {
              "title": "X-Contract-Version",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveReportVersionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportDefinitionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Approve Version"
      }
    },
    "/reports/definitions/{definition_id}/versions/{version}/blocks": {
      "put": {
        "operationId": "reportReplaceDraftBlocks",
        "parameters": [
          {
            "in": "path",
            "name": "definition_id",
            "required": true,
            "schema": {
              "title": "Definition Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "version",
            "required": true,
            "schema": {
              "title": "Version",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-As-Of",
            "required": true,
            "schema": {
              "title": "X-As-Of",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Trace-Id",
            "required": true,
            "schema": {
              "title": "X-Trace-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Timezone",
            "required": true,
            "schema": {
              "title": "X-Timezone",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Contract-Version",
            "required": true,
            "schema": {
              "title": "X-Contract-Version",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplaceReportBlocksRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportDefinitionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Replace Draft Blocks"
      }
    },
    "/reports/definitions/{definition_id}/versions/{version}/drafts": {
      "post": {
        "operationId": "reportCreateNextDraft",
        "parameters": [
          {
            "in": "path",
            "name": "definition_id",
            "required": true,
            "schema": {
              "title": "Definition Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "version",
            "required": true,
            "schema": {
              "title": "Version",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-As-Of",
            "required": true,
            "schema": {
              "title": "X-As-Of",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Trace-Id",
            "required": true,
            "schema": {
              "title": "X-Trace-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Timezone",
            "required": true,
            "schema": {
              "title": "X-Timezone",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Contract-Version",
            "required": true,
            "schema": {
              "title": "X-Contract-Version",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportDefinitionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Create Next Draft"
      }
    },
    "/reports/runs": {
      "get": {
        "operationId": "reportListRuns",
        "parameters": [
          {
            "in": "query",
            "name": "definition_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Definition Id"
            }
          },
          {
            "in": "header",
            "name": "X-As-Of",
            "required": true,
            "schema": {
              "title": "X-As-Of",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Trace-Id",
            "required": true,
            "schema": {
              "title": "X-Trace-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Timezone",
            "required": true,
            "schema": {
              "title": "X-Timezone",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Contract-Version",
            "required": true,
            "schema": {
              "title": "X-Contract-Version",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportRunListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "List Runs"
      }
    },
    "/reports/runs/manual": {
      "post": {
        "operationId": "reportCreateManualRunCommand",
        "parameters": [
          {
            "in": "header",
            "name": "X-As-Of",
            "required": true,
            "schema": {
              "title": "X-As-Of",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Trace-Id",
            "required": true,
            "schema": {
              "title": "X-Trace-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Timezone",
            "required": true,
            "schema": {
              "title": "X-Timezone",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Contract-Version",
            "required": true,
            "schema": {
              "title": "X-Contract-Version",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateManualRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManualRunCommandResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Create Manual Run Command"
      }
    },
...
  }
}